refactoring unit model moved UnitLookup to unitmodel.py
