+++ qps/layerconfigwidgets/rasterbands.py
@@ -42,8 +42,9 @@ from qgis.core import QgsRasterLayer, QgsMapLayer, \
 from qgis.gui import QgsMapCanvas, QgsMapLayerConfigWidget, QgsMapLayerConfigWidgetFactory, QgsRasterBandComboBox
 from ..layerconfigwidgets.core import QpsMapLayerConfigWidget
 from ..simplewidgets import FlowLayout
-from ..utils import loadUi, parseWavelength, UnitLookup, parseFWHM, LUT_WAVELENGTH, WAVELENGTH_DESCRIPTION, \
+from ..utils import loadUi, parseWavelength, parseFWHM, LUT_WAVELENGTH, WAVELENGTH_DESCRIPTION, \
     SignalBlocker, printCaller, rendererXML
+from ..unitmodel import UnitLookup
 
 
 class RasterBandComboBox(QgsRasterBandComboBox):
