+++ qps/testing.py
@@ -71,7 +71,8 @@ from .speclib import createStandardFields, FIELD_VALUES
 from .speclib.core import profile_fields as pFields, create_profile_field, is_profile_field, profile_field_indices
 from .speclib.core.spectrallibrary import SpectralLibraryUtils
 from .speclib.core.spectralprofile import prepareProfileValueDict, encodeProfileValueDict
-from .utils import UnitLookup, px2geo, SpatialPoint, findUpwardPath
+from .utils import px2geo, SpatialPoint, findUpwardPath
+from .unitmodel import UnitLookup
 
 if Qgis.versionInt() >= 33000:
     TYPE_WkbType = Qgis.WkbType
