+++ qps/unitmodel.py
@@ -1,13 +1,22 @@
-from typing import Dict
+import calendar
+import copy
+import datetime
+import re
+import warnings
+from math import log10
+from typing import Dict, List, Union
 
 import numpy as np
-
-from qgis.PyQt.QtCore import Qt, NULL, QAbstractListModel, QModelIndex
+from qgis.PyQt.QtCore import Qt, NULL, QAbstractListModel, QModelIndex, QDate, QDateTime
 from qgis.PyQt.QtGui import QIcon
-from .utils import UnitLookup, METRIC_EXPONENTS, datetime64
 
 BAND_INDEX = 'Band Index'
 BAND_NUMBER = 'Band Number'
+UNKNOWN = ''
+
+METRIC_EXPONENTS = {
+    'nm': -9, 'μm': -6, 'mm': -3, 'cm': -2, 'dm': -1, 'm': 0, 'hm': 2, 'km': 3
+}
 
 
 class UnitModel(QAbstractListModel):
@@ -16,6 +25,10 @@ class UnitModel(QAbstractListModel):
         super().__init__(*args, **kwds)
 
         self.mUnits: Dict[str, tuple] = dict()
+        self.mRedundantUnits: bool = False
+
+    def setAllowRedundantUnits(self, b: bool):
+        self.mRedundantUnits = b
 
     def setAllowEmptyUnit(self,
                           allowEmpty: bool,
@@ -223,7 +236,437 @@ class XUnitModel(UnitModel):
         self.addUnit('DecimalYear', description='Decimal Year')
         self.addUnit('DOY', description='Day of Year')
 
+        for u in ['Meters', 'Kilometers', 'Yards', 'Miles']:
+            baseUnit = UnitLookup.baseUnit(u)
+            assert isinstance(baseUnit, str), u
+            self.addUnit(baseUnit, description='{} []')
+
+        self.addUnit('')
+        self.addUnit(UNKNOWN, description='Unknown Unit')
+
     def findUnit(self, unit):
         if unit in [None, NULL]:
             unit = BAND_NUMBER
         return super(XUnitModel, self).findUnit(unit)
+
+
+def square_with_sign(v):
+    if v >= 0:
+        return v ** 2
+    else:
+        return -(v ** 2)
+
+
+def log10_with_sign(v):
+    if v >= 0:
+        return log10(v)
+    else:
+        return -(log10(abs(v)))
+
+
+class UnitLookup(object):
+    DATE_UNITS = ['DateTime', 'DOY', 'DecimalYear', 'DecimalYear[366]', 'DecimalYear[365]', 'Y', 'M', 'W', 'D']
+    TIME_UNITS = ['h', 'm', 's', 'ms', 'us', 'ns', 'ps', 'fs', 'as']
+
+    IMPERIAL_LENGTH_UNITS = {
+        # values in m (si meters)
+        'in': 0.0254, 'ft': 0.3048, 'yd': 0.9144, 'mi': 1609.344, 'nmi': 1852
+    }
+
+    IMPERIAL_AREA_UNITS = {
+        # values in sqm (si square meters)
+        'acre': 4046.8564224, 'sqmi': 2589988.110336
+    }
+
+    # Length units with exponents related to 1 SI meter [m]
+    LENGTH_UNITS = METRIC_EXPONENTS.copy()
+    LENGTH_UNITS.update({k: log10(v) for k, v in IMPERIAL_LENGTH_UNITS.items()})
+
+    # Area units with exponents related to 1 SI square meter [m²]
+    # AREA_UNITS = {f'{k}²': square_with_sign(v) for k, v in LENGTH_UNITS.items()}
+    # AREA_UNITS.update({f'{k}': log10_with_sign(v) for k, v in IMPERIAL_AREA_UNITS.items()})
+    AREA_UNITS = {f'{k}²': 2 * v for k, v in LENGTH_UNITS.items()}
+    AREA_UNITS['ha'] = 4  # because 10^4 m = 1 ha
+    AREA_UNITS.update({f'{k}': 2 * v for k, v in IMPERIAL_AREA_UNITS.items()})
+
+    # a dictionary to lookup other names of length or area units
+    # e.g. UNIT_LOOKUP['meters'] = 'm'
+    # will be filled dynamically
+    UNIT_LOOKUP = {}
+
+    @staticmethod
+    def metric_units() -> List[str]:
+        warnings.warn(DeprecationWarning('Use area_units() or length_units()'))
+        return list(METRIC_EXPONENTS.keys())
+
+    @staticmethod
+    def length_units() -> List[str]:
+        return list(UnitLookup.LENGTH_UNITS.keys())
+
+    @staticmethod
+    def area_units() -> List[str]:
+        return list(UnitLookup.AREA_UNITS.keys())
+
+    @staticmethod
+    def date_units() -> List[str]:
+        return list(UnitLookup.DATE_UNITS)
+
+    @staticmethod
+    def time_units() -> List[str]:
+        return list(UnitLookup.TIME_UNITS)
+
+    @staticmethod
+    def baseUnit(unit: str) -> str:
+        """
+        Tries to return the basic physical unit
+        e.g. "m" for string of "Meters"
+
+        :param unit:
+        :type unit:
+        :return:
+        :rtype:
+        """
+        if not isinstance(unit, str):
+            return None
+
+        unit = unit.strip()
+
+        if unit in UnitLookup.UNIT_LOOKUP.keys():
+            return UnitLookup.UNIT_LOOKUP[unit]
+
+        # so far this unit is unknown. Try to find the base unit
+        # store unit string in Lookup table for fast conversion into its base unit
+        # e.g. to convert string like "MiKrOMetErS" to "μm"
+        base_unit = None
+
+        if unit in UnitLookup.length_units() + \
+                UnitLookup.area_units() + \
+                UnitLookup.date_units() + \
+                UnitLookup.time_units():
+            base_unit = unit
+        elif re.search(r'^(Nanomet(er|re)s?)$', unit, re.I):
+            base_unit = 'nm'
+        elif re.search(r'^(Micromet(er|re)s?|um|μm)$', unit, re.I):
+            base_unit = 'μm'
+        elif re.search(r'^(Millimet(er|re)s?)$', unit, re.I):
+            base_unit = 'mm'
+        elif re.search(r'^(Centimet(er|re)s?)$', unit, re.I):
+            base_unit = 'cm'
+        elif re.search(r'^(Decimet(er|re)s?)$', unit, re.I):
+            base_unit = 'dm'
+        elif re.search(r'^(Met(er|re)s?)$', unit, re.I):
+            base_unit = 'm'
+        elif re.search(r'^(Hectomet(er|re)s?)$', unit, re.I):
+            base_unit = 'hm'
+        elif re.search(r'^(Kilomet(er|re)s?)$', unit, re.I):
+            base_unit = 'km'
+        # date units
+        elif re.search(r'(Date([_\- ]?Time)?([_\- ]?Group)?|DTG)$', unit, re.I):
+            base_unit = 'DateTime'
+        elif re.search(r'^(doy|Day[-_ ]?Of[-_ ]?Year?)$', unit, re.I):
+            base_unit = 'DOY'
+        elif re.search(r'decimal[_\- ]?years?$', unit, re.I):
+            base_unit = 'DecimalYear'
+        elif re.search(r'decimal[_\- ]?years?\[356\]$', unit, re.I):
+            base_unit = 'DecimalYear[365]'
+        elif re.search(r'decimal[_\- ]?years?\[366\]$', unit, re.I):
+            base_unit = 'DecimalYear[366]'
+        elif re.search(r'^Years?$', unit, re.I):
+            base_unit = 'Y'
+        elif re.search(r'^Months?$', unit, re.I):
+            base_unit = 'M'
+        elif re.search(r'^Weeks?$', unit, re.I):
+            base_unit = 'W'
+        elif re.search(r'^Days?$', unit, re.I):
+            base_unit = 'D'
+        elif re.search(r'^Hours?$', unit, re.I):
+            base_unit = 'h'
+        elif re.search(r'^Minutes?$', unit, re.I):
+            base_unit = 'm'
+        elif re.search(r'^Seconds?$', unit, re.I):
+            base_unit = 's'
+        elif re.search(r'^MilliSeconds?$', unit, re.I):
+            base_unit = 'ms'
+        elif re.search(r'^MicroSeconds?$', unit, re.I):
+            base_unit = 'us'
+        elif re.search(r'^NanoSeconds?$', unit, re.I):
+            base_unit = 'ns'
+        elif re.search(r'^Picoseconds?$', unit, re.I):
+            base_unit = 'ps'
+        elif re.search(r'^Femtoseconds?$', unit, re.I):
+            base_unit = 'fs'
+        elif re.search(r'^Attoseconds?$', unit, re.I):
+            base_unit = 'as'
+
+        # imperial length units
+        elif re.search(r'^inch(es)?$', unit, re.I):
+            base_unit = 'in'
+        elif re.search(r'^foot$', unit, re.I):
+            base_unit = 'ft'
+        elif re.search(r'^yards?$', unit, re.I):
+            base_unit = 'yd'
+        elif re.search(r'^miles?$', unit, re.I):
+            base_unit = 'mi'
+        elif re.search(r'^nautical[_ ]miles?$', unit, re.I):
+            base_unit = 'nmi'
+
+        # imperial area units
+        elif re.search(r'^acres?$', unit, re.I):
+            base_unit = 'acre'
+        elif re.search(r'^square miles?$', unit, re.I):
+            base_unit = 'sqmi'
+
+        if base_unit:
+            UnitLookup.UNIT_LOOKUP[unit] = base_unit
+        return base_unit
+
+    @staticmethod
+    def isMetricUnit(unit: str) -> bool:
+        baseUnit = UnitLookup.baseUnit(unit)
+        return baseUnit in UnitLookup.metric_units()
+
+    @staticmethod
+    def isTemporalUnit(unit: str) -> bool:
+        baseUnit = UnitLookup.baseUnit(unit)
+        return baseUnit in UnitLookup.time_units() + UnitLookup.date_units()
+
+    @staticmethod
+    def convertMetricUnit(value: Union[float, np.ndarray], u1: str, u2: str) -> float:
+        """
+        Converts value `value` from unit `u1` into unit `u2`
+        :param value: float | int | might work with numpy arrays as well
+        :param u1: str, identifier of unit 1
+        :param u2: str, identifier of unit 2
+        :return: float | numpy.array, converted values
+                 or None in case conversion is not possible
+        """
+        assert isinstance(u1, str), 'Source unit needs to be a str'
+        assert isinstance(u2, str), 'Destination unit needs to be a str'
+        u1 = UnitLookup.baseUnit(u1)
+        u2 = UnitLookup.baseUnit(u2)
+
+        e1 = UnitLookup.METRIC_EXPONENTS.get(u1)
+        e2 = UnitLookup.METRIC_EXPONENTS.get(u2)
+
+        if all([arg is not None for arg in [value, e1, e2]]):
+            if e1 == e2:
+                return copy.copy(value)
+            elif isinstance(value, list):
+                return [v * 10 ** (e1 - e2) for v in value]
+            else:
+                return value * 10 ** (e1 - e2)
+        else:
+            return None
+
+    @staticmethod
+    def convertLengthUnit(value: Union[float, np.ndarray], u1: str, u2: str) -> float:
+        """
+        Converts a length value `value` from unit `u1` into unit `u2`
+        :param value: float | int | might work with numpy arrays as well
+        :param u1: str, identifier of unit 1
+        :param u2: str, identifier of unit 2
+        :return: float | numpy.array, converted values
+                 or None in case conversion is not possible
+        """
+        assert isinstance(u1, str), 'Source length unit (str)'
+        assert isinstance(u2, str), 'Destination length unit (str)'
+        u1 = UnitLookup.baseUnit(u1)
+        u2 = UnitLookup.baseUnit(u2)
+
+        # get exponents to convert from unit to meters [m]
+        e1 = UnitLookup.LENGTH_UNITS.get(u1)
+        e2 = UnitLookup.LENGTH_UNITS.get(u2)
+
+        if all([arg is not None for arg in [value, e1, e2]]):
+            if e1 == e2:
+                return copy.copy(value)
+            elif isinstance(value, list):
+                return [v * 10 ** (e1 - e2) for v in value]
+            else:
+                return value * 10 ** (e1 - e2)
+        else:
+            return None
+
+    @staticmethod
+    def convertAreaUnit(value: Union[float, np.ndarray], u1: str, u2: str) -> float:
+        """
+        Converts an area value `value` from unit `u1` into unit `u2`
+        :param value: float | int | might work with numpy arrays as well
+        :param u1: str, identifier of unit 1
+        :param u2: str, identifier of unit 2
+        :return: float | numpy.array, converted values
+                 or None in case conversion is not possible
+        """
+        assert isinstance(u1, str), 'Source area unit (str)'
+        assert isinstance(u2, str), 'Destination area unit (str)'
+
+        # get the base unit, e.g. 'm2' instead 'square meters'
+        u1 = UnitLookup.baseUnit(u1)
+        u2 = UnitLookup.baseUnit(u2)
+
+        # get exponents to convert from unit to square meters [m²]
+        e1 = UnitLookup.AREA_UNITS.get(u1)
+        e2 = UnitLookup.AREA_UNITS.get(u2)
+
+        if all([arg is not None for arg in [value, e1, e2]]):
+            if e1 == e2:
+                return copy.copy(value)
+            elif isinstance(value, list):
+                return [v * 10 ** (e1 - e2) for v in value]
+            else:
+                return value * 10 ** (e1 - e2)
+        else:
+            return None
+
+    @staticmethod
+    def convertDateUnit(value: np.datetime64, unit: str):
+        """
+        Converts a
+        :param value: numpy.datetime64 | datetime.date | datetime.datetime | float | int
+                      int values are interpreted as year
+                      float values are interpreted as decimal year
+        :param unit: output unit
+                    (integer) Y - Year, M - Month, W - Week, D - Day, DOY - Day-of-Year
+                    (float) DecimalYear (based on True number of days per year)
+                    (float) DecimalYear[365] (based on 365 days per year, i.e. wrong for leap years)
+                    (float) DecimalYear[366] (based on 366 days per year, i.e. wrong for none-leap years)
+
+        :return: float (if unit is decimal year), int else
+        """
+        unit = UnitLookup.baseUnit(unit)
+        if not UnitLookup.isTemporalUnit(unit):
+            return None
+        # see https://numpy.org/doc/stable/reference/arrays.datetime.html#arrays-dtypes-dateunits
+        # for valid date units
+        if isinstance(value, (np.ndarray, list)):
+            func = np.vectorize(UnitLookup.convertDateUnit)
+            return func(value, unit)
+
+        value = datetime64(value)
+        if unit == 'Y':
+            return value.astype(object).year
+        elif unit == 'M':
+            return value.astype(object).month
+        elif unit == 'D':
+            return value.astype(object).day
+        elif unit == 'W':
+            return value.astype(object).week
+        elif unit == 'DOY':
+            return ((value - value.astype('datetime64[Y]')).astype('timedelta64[D]') + 1).astype(int)
+
+        elif unit.startswith('DecimalYear'):
+            year = value.astype(object).year
+            year64 = value.astype('datetime64[Y]')
+
+            # second of year
+            soy = (value - year64).astype('timedelta64[s]').astype(np.float64)
+
+            # seconds per year
+            if unit == 'DecimalYear[366]':
+                spy = 366 * 86400
+            elif unit == 'DecimalYear[365]':
+                spy = 365 * 86400
+            else:
+                spy = 366 if calendar.isleap(year) else 365
+                spy *= 86400
+            spy2 = np.datetime64('{:04}-01-01T00:00:00'.format(year + 1)) - np.datetime64(
+                '{:04}-01-01T00:00:00'.format(year))
+            spy2 = int(spy2.astype(int))
+            if spy != spy2:
+                s = ""
+            return float(year + soy / spy)
+        else:
+            raise NotImplementedError()
+
+
+def datetime64(value, dpy: int = None) -> np.datetime64:
+    """
+    Converts an input value into a numpy.datetime64 value.
+    :param value: the value to be converted into a numpy.datetime64 value
+    :param dpy: days per year. If `value` is a float, it is considered to be a decimal year value.
+                    By default it is assumed that the year fraction is calculated on 366 year in leap years and 365
+                    in none-leap year. However, dpy can be used to use any other number of days per year to convert
+                    the fraction back into days.
+    :return: numpy.datetime64
+    """
+    if isinstance(value, np.datetime64):
+        return value
+    elif isinstance(value, QDate):
+        return np.datetime64(value.toPyDate())
+    elif isinstance(value, QDateTime):
+        return np.datetime64(value.toPyDateTime())
+    elif isinstance(value, (str, datetime.date, datetime.datetime)):
+        return np.datetime64(value)
+    elif isinstance(value, int):
+        # expect a year
+        return np.datetime64('{:04}-01-01'.format(value))
+    elif isinstance(value, float):
+        # expect a decimal year
+        year = int(value)
+        fraction = value - year
+
+        if dpy is None:
+            dpy = 366 if calendar.isleap(year) else 365
+        else:
+            assert dpy in [365, 366]
+        # seconds of year
+        soy = np.round(fraction * dpy * 86400).astype(int)
+        return np.datetime64('{:04}-01-01'.format(year)) + np.timedelta64(soy, 's')
+
+    if isinstance(value, np.ndarray):
+        func = np.vectorize(datetime64)
+        return func(value)
+    elif isinstance(value, list):
+        return datetime64(np.asarray(value), dpy=dpy)
+    else:
+        raise NotImplementedError('Unsupported input value: {}'.format(value))
+
+
+def day_of_year(date: np.datetime64) -> int:
+    """
+    Returns a date's Day-of-Year (DOY) (considering leap-years)
+    :param date: numpy.datetime64
+    :return: numpy.ndarray[int]
+    """
+    if not isinstance(date, np.datetime64):
+        date = np.datetime64(date)
+
+    dt = date - date.astype('datetime64[Y]') + 1
+    return dt.astype(int)
+
+
+def days_per_year(year):
+    """
+    Returns the days per year
+    :param year:
+    :return:
+    """
+    # is it a leap year?
+    if isinstance(year, float):
+        year = int(year)
+    elif isinstance(year, np.number):
+        year = int(year)
+    elif isinstance(year, np.datetime64):
+        year = year.astype(object).year
+    elif isinstance(year, datetime.date):
+        year = year.year
+    elif isinstance(year, datetime.datetime):
+        year = year.year
+    elif isinstance(year, np.ndarray):
+        func = np.vectorize(days_per_year)
+        return func(year)
+
+    return 366 if calendar.isleap(year) else 365
+
+    """
+    1. If the year is evenly divisible by 4, go to step 2. Otherwise, False.
+    2. If the year is evenly divisible by 100, go to step 3. Otherwise, False
+    3. If the year is evenly divisible by 400, True Otherwise, False
+
+    """
+    """
+    Every year that is exactly divisible by four is a leap year, except for years that are exactly divisible by 100,
+    but these centurial years are leap years, if they are exactly divisible by 400.
+    """
+    # is_leap = (year % 4 == 0 and not year % 100 == 0) or (year % 100 == 0 and year % 400 == 0)
+    # return np.where(is_leap, 366, 365)
